requirements.txt
Flask
flask-cors
scikit-learn
opencv-python
numpy
Pillow
gunicorn
